Change search box to store SamwiseUserProfile type
